+++ src/main.js
@@ -2,16 +2,35 @@ import SiteMenuView from './view/site-menu.js';
 import StatisticsView from './view/statistics.js';
 import ProfileView from './view/profile.js';
 import {generateFilm} from './mock/film.js';
+import {generateComments} from './mock/comments.js';
 import {getNumberFilms} from './utils/films.js';
 import {render} from './utils/dom.js';
 import FilmBoardPresenter from './presenter/film-board.js';
 import FilmsModel from './model/films.js';
 import CommentsModel from './model/comments.js';
 import {generateFilter} from './utils/filter.js';
+import {getRandomInteger} from './utils/common.js';
+
+const MIN_COMMENTS_COUNT = 0;
+const MAX_COMMENTS_COUNT = 7;
 
 const FILM_DEVELOPER_COUNT = 22;
 
 const films = new Array(FILM_DEVELOPER_COUNT).fill().map(() => generateFilm());
+let comments = [];
+
+films.forEach((film) => {
+  const currenComments = generateComments(getRandomInteger(MIN_COMMENTS_COUNT, MAX_COMMENTS_COUNT), film.id);
+
+  currenComments.forEach((comment) => {
+    if (comment.filmId === film.id) {
+      film.comments.push(comment.id);
+    }
+  });
+
+  comments = comments.concat(currenComments);
+});
+
 const numberFilms = getNumberFilms(films);
 
 const filmModel = new FilmsModel();
