+++ src/mock/film.js
@@ -1,10 +1,6 @@
-import {EMOJI} from '../const.js';
 import {nanoid} from 'nanoid';
 import {getRandomInteger, getRandomFloat} from '../utils/common.js';
-
-const MAX_COMMENTS_COUNT = 50;
 const MAX_DESCRIPTION_COUNT = 5;
-const MIN_COMMENTS_COUNT = 0;
 const MIN_DESCRIPTION_COUNT = 1;
 
 const getRandomBoolean = () => Boolean(getRandomInteger(0, 1));
@@ -70,37 +66,7 @@ const generateGenres = (count) => {
   return new Array(count).fill().map(() => getRandomArrayItem(GENRES));
 };
 
-const generateEmoji = () => EMOJI[getRandomInteger(0, EMOJI.length - 1)];
-
-const generateText = () => {
-  const TEXTS = [
-    'Interesting setting and a good cast',
-    'Booooooooooring',
-    'Very very old. Meh',
-    'Almost two hours? Seriously?',
-  ];
-
-  return getRandomArrayItem(TEXTS);
-};
-
-const generateAuthor = () => {
-  const AUTHORS = [
-    'Tim Macoveev',
-    'John Doe',
-  ];
-
-  return getRandomArrayItem(AUTHORS);
-};
-
 const generateDate = () => `${getRandomInteger(1950, 2021)}-${getRandomInteger(1, 12)}-${getRandomInteger(1, 28)}`;
-
-const generateComments = (count) => new Array(count).fill().map(() => ({
-  id: nanoid(),
-  text: generateText(),
-  emoji: generateEmoji(),
-  author: generateAuthor(),
-  date: generateDate(),
-}));
 
 const generateFilm = () => ({
   id: nanoid(),
@@ -108,7 +74,7 @@ const generateFilm = () => ({
   original: generateFilmName(),
   img: generateFilmImage(),
   description: generateDescription(getRandomInteger(MIN_DESCRIPTION_COUNT, MAX_DESCRIPTION_COUNT)),
-  comments: generateComments(getRandomInteger(MIN_COMMENTS_COUNT, MAX_COMMENTS_COUNT)),
+  comments: [],
   rating: getRandomFloat(0, 10, 1),
   release: generateDate(),
   duration: getRandomInteger(0, 220),
